% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Annotate.PolyA}
\alias{Annotate.PolyA}
\title{Annotate the PACs}
\usage{
Annotate.PolyA(QpolyA,gff,seq.levels=NA)
}
\arguments{
\item{QpolyA}{A QuantifyPolyA object containing clean poly(A) sites and PACs.}

\item{gff}{A genome annotation file in GFF or GTF format, GTF format is recommended.}

\item{seq.levels}{A character vector of the names of Chromosomes/Contigs. This parameter is useful when the Chromosomes/Contigs names in genome file (used for short reads mapping) and genome annotation file are not consistent.}
}
\description{
Annotate the PACs based on existed genome annotations, and generate the PAC table of all experimental samples.
}
