% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{check_pt_tags}
\alias{check_pt_tags}
\title{Check if SAM file contains pt tags by sampling lines}
\usage{
check_pt_tags(sam_file_path, sample_size = 1000)
}
\arguments{
\item{sam_file_path}{Path to SAM file}

\item{sample_size}{Number of lines to sample}
}
\value{
TRUE if pt tags found, FALSE otherwise
}
\description{
Check if SAM file contains pt tags by sampling lines
}
