% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Cluster.PolyA}
\alias{Cluster.PolyA}
\title{Generate PACs}
\usage{
Cluster.PolyA(QpolyA, max.gapwidth = 24, mc.cores = 4)
}
\arguments{
\item{QpolyA}{A QuantifyPolyA object containing clean poly(A) sites.}

\item{max.gapwidth}{A cutoff limiting the max distance between two adjacent poly(A) sites in a poly(A) site cluster (PAC), default value is 24.}

\item{mc.cores}{An integer indicating the number of processors/cores to perform the clustering, default value is 4.}
}
\description{
Cluster poly(A) sites into clusters using a weighted density peak clustering algorithm.
}
