% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{DESeq2.PolyA}
\alias{DESeq2.PolyA}
\title{Perform basic differential expression analysis using DESeq2 package}
\usage{
DESeq2.PolyA(QpolyA,colData)
}
\arguments{
\item{QpolyA}{A QuantifyPolyA object.}

\item{colData}{A data.frame object containing the information of experimental design.}
}
\description{
Perform basic differential expression analysis using DESeq2 package
}
