% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{dynamicsDetect}
\alias{dynamicsDetect}
\title{Detect APA dynamics between two samples}
\usage{
dynamicsDetect(sample_x, sample_y, sample_strand, sample_center)
}
\arguments{
\item{sample_x}{A data.frame object with multiple rows and columns, each row represents a PAC, each column represents a replicate.}

\item{sample_y}{A data.frame object with multiple rows and columns, each row represents a PAC, each column represents a replicate.}

\item{sample_strand}{A vector containing the strand information of PACs.}

\item{sample_center}{A vector containing the coordinates (centers) information of PACs.}
}
\value{
A data.frame object containing three columns 'pd', 'r', 'p.value'.
}
\description{
Detect APA dynamics between two samples
}
