% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Extract_polyAsite}
\alias{Extract_polyAsite}
\title{Process SAM files to detect polyA tails and generate processed BED files}
\usage{
Extract_polyAsite(
  work_dir,
  intron_max = 50000,
  min_tail_length = 6,
  sample_size = 1000,
  bedtools_path = "bedtools",
  remove_temp_files = FALSE
)
}
\arguments{
\item{work_dir}{Working directory containing SAM files}

\item{intron_max}{Maximum intron size, default is 50000}

\item{min_tail_length}{Minimum polyA tail length to consider, default is 6}

\item{sample_size}{Number of lines to sample for pt tag detection, default is 1000}

\item{bedtools_path}{Path to bedtools executable, default is "bedtools"}

\item{remove_temp_files}{Whether to remove temporary files, default is FALSE}
}
\value{
List containing processing statistics for each file
}
\description{
Process SAM files to detect polyA tails and generate processed BED files
}
