% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Filter.PolyA}
\alias{Filter.PolyA}
\title{Filter out PACs}
\usage{
Filter.PolyA(QpolyA,min_count = 10, min_sample = 1)
}
\arguments{
\item{QpolyA}{A QuantifyPolyA object containing clean poly(A) sites and PAC table.}

\item{min_count}{A fixed value specifying the minimum read count in a PAC.}

\item{min_sample}{A fixed value specifying the minimum number of samples with min_count reads in a PAC.}
}
\description{
Filter out PACs with low count of supporting reads across multiple samples.
}
