% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{FU}
\alias{FU}
\title{Calculate the Pearson correlation coefficient of two PAC vectors.}
\usage{
FU(s1, s2, score)
}
\arguments{
\item{s1}{A numeric vector of PAC usages of sample X.}

\item{s2}{A numeric vector of PAC usages of sample Y.}

\item{score}{A numeric vector containing the genomic coordinates of PACs.}
}
\value{
A numeric value between -1 and 1.
}
\description{
Calculate the Pearson correlation coefficient of two PAC vectors.
}
