% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{is.internal.priming}
\alias{is.internal.priming}
\title{Check whether the poly(A) sites are resulted from internal priming artifacts or not.}
\usage{
is.internal.priming(polyA, fasta, flank_len = 15, win_size = 10, min_A = 8)
}
\arguments{
\item{polyA}{A data.frame containing the information of raw poly(A) sites.}

\item{fasta}{A string specifies the location and name of genome file in FASTA format.}

\item{flank_len}{An integer specifying the neighborhood to search for template polyA.}

\item{win_size}{An integer specifying the size of the sliding window.}

\item{min_A}{An integer specifying the minimum number of base A in a window.}
}
\value{
A logic vector indicates whether the poly(A) sites are resulted from internal priming artifacts or not.
}
\description{
Check whether the poly(A) sites are resulted from internal priming artifacts or not.
}
