% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Load.PolyA}
\alias{Load.PolyA}
\title{Load raw poly(A) site data}
\usage{
Load.PolyA(files,dir)
}
\arguments{
\item{files}{A character vector specifying the names of files containing poly(A) site information, each file contains 6 columns: 'seqnames','strand','coord','score','five_prime_end','tail_lengths'.}

\item{dir}{A string specifying the directory of files containing poly(A) site information.}
}
\value{
A QuantifyPolyA object containing all raw poly(A) site information.
}
\description{
Load the raw poly(A) site information extracted from 3' end sequencing data.
}
