% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Map.Tail}
\alias{Map.Tail}
\title{Map tail lengths to clusters}
\usage{
Map.Tail(QpolyA, delimiter = ";")
}
\arguments{
\item{QpolyA}{A QuantifyPolyA object containing clean poly(A) sites and PACs.}

\item{delimiter}{The delimiter to use for tail lengths (default: ";")}
}
\value{
A QuantifyPolyA object with tail length information mapped to clusters.
}
\description{
Map poly(A) tail length information to PAS clusters
}
