% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{minimap2}
\alias{minimap2}
\title{Batch align fq files using minimap2 and filter/sort SAM files}
\usage{
minimap2(reference, work_dir, threads = 4, filter_flags = 2308)
}
\arguments{
\item{reference}{Reference genome file path}

\item{work_dir}{Working directory path containing fq files}

\item{threads}{Number of threads for minimap2, default is 4}

\item{filter_flags}{SAM flags to filter out, default is 2308}
}
\value{
Vector of generated filtered SAM file paths
}
\description{
Batch align fq files using minimap2 and filter/sort SAM files
}
