% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Motif.Search}
\alias{Motif.Search}
\title{Motif search}
\usage{
Motif.Search(Sites, fasta)
}
\arguments{
\item{Sites}{A data.frame containing the information of poly(A) sites.}

\item{fasta}{A string specifies the location and name of genome file in FASTA format.}
}
\value{
A list containing the types and genomic locations of signals.
}
\description{
Search canonical poly(A) signals at the upstream of poly(A) sites.
}
