% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{perform_wilcoxon_all_mRNA_multi}
\alias{perform_wilcoxon_all_mRNA_multi}
\title{Perform Wilcoxon test for multiple treatments vs NC}
\usage{
perform_wilcoxon_all_mRNA_multi(
  cluster_data,
  logscale = TRUE,
  control_group = "NC"
)
}
\arguments{
\item{cluster_data}{Data for a single cluster}

\item{logscale}{Whether to log2 transform tail lengths}

\item{control_group}{Name of control group (default: "NC")}
}
\description{
Perform Wilcoxon test for multiple treatments vs NC
}
