% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Plot.TailDensity}
\alias{Plot.TailDensity}
\title{Plot poly(A) tail-length density}
\usage{
Plot.TailDensity(
  x,
  colors = c(NC = "#E64B35", EX1 = "#4DBBD5", EX2 = "#00A087")
)
}
\arguments{
\item{x}{A result returned by \code{\link[=Tail.PCA]{Tail.PCA()}}.}

\item{colors}{Named colors for sample conditions.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot poly(A) tail-length density
}
