% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Plot.TailVolcano}
\alias{Plot.TailVolcano}
\title{Plot differential poly(A) tail lengths}
\usage{
Plot.TailVolcano(
  result,
  title,
  control_group = "NC",
  q_cutoff = 0.05,
  mean_diff_cutoff = 15
)
}
\arguments{
\item{result}{A result returned by \code{\link[=Tail.DiffPair]{Tail.DiffPair()}}.}

\item{title}{Plot title.}

\item{control_group}{Control-group label.}

\item{q_cutoff}{Adjusted P-value cutoff.}

\item{mean_diff_cutoff}{Absolute mean-difference cutoff in nucleotides.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot differential poly(A) tail lengths
}
