% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{polyAsite.annotation}
\alias{polyAsite.annotation}
\title{Annotate the poly(A) sites based on the genomic structures}
\usage{
polyAsite.annotation(polyA, gff, seq.levels=NA)
}
\arguments{
\item{polyA}{A data.frame containing the info of clustered poly(A) sites - PAC.}

\item{gff}{A genome annotation file in GFF or GTF format, GTF format is recommended.}

\item{seq.levels}{A character vector of the names of Chromosomes/Contigs. This parameter is useful when the Chromosomes/Contigs names in genome file (used for short reads mapping) and genome annotation file are not consistent.}
}
\value{
The polyA data.frame will be return with three additional columns 'gene_id', 'distance, 'type'.
}
\description{
Annotate the poly(A) sites based on the genomic structures
}
