% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{prepare_cluster_tail_data_fast}
\alias{prepare_cluster_tail_data_fast}
\title{Convert the raw tail length data into a long table}
\usage{
prepare_cluster_tail_data_fast(
  QpolyA,
  sample_info,
  cores = 4,
  show_progress = TRUE
)
}
\arguments{
\item{QpolyA}{A QuantifyPolyA object}

\item{sample_info}{Data frame with sample metadata}

\item{cores}{Number of CPU cores to use}

\item{show_progress}{Whether to show processing progress.}
}
\value{
A data table with individual mRNA tail lengths (cluster_id, sample, condition, tail_length, lib_id)
}
\description{
Fast parallel processing with only essential columns
}
