% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{prepare_tail_length_matrix}
\alias{prepare_tail_length_matrix}
\title{Prepare tail length matrix for PCA}
\usage{
prepare_tail_length_matrix(QpolyA, sample_info, aggregation_method = "mean")
}
\arguments{
\item{QpolyA}{QuantifyPolyA object}

\item{sample_info}{Sample metadata}

\item{aggregation_method}{Method to aggregate tail lengths: "mean", "median"}
}
\value{
A matrix for PCA analysis
}
\description{
Create sample x PAS cluster matrix with aggregated tail lengths
}
