% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{print_summary}
\alias{print_summary}
\title{Print summary of processing results}
\usage{
print_summary(results_list)
}
\arguments{
\item{results_list}{List containing processing statistics for all files}
}
\description{
Print summary of processing results
}
