% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{process_single_sam}
\alias{process_single_sam}
\title{Process a single SAM file and generate BED output}
\usage{
process_single_sam(
  sam_file_path,
  bed_file_path,
  intron_max,
  min_tail_length,
  sample_size
)
}
\arguments{
\item{sam_file_path}{Path to input SAM file}

\item{bed_file_path}{Path to output BED file}

\item{intron_max}{Maximum intron size}

\item{min_tail_length}{Minimum polyA tail length}

\item{sample_size}{Number of lines to sample for pt tag detection}
}
\value{
List containing processing statistics
}
\description{
Process a single SAM file and generate BED output
}
