% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Quantify.CanonicalAPA}
\alias{Quantify.CanonicalAPA}
\title{Quantify APA dynamics among canonical PACs}
\usage{
Quantify.CanonicalAPA(QpolyA,colData,contrast)
}
\arguments{
\item{QpolyA}{A QuantifyPolyA object generated after poly(A) site clustering and annotation.}

\item{colData}{A data.frame object containing the information of experimental design.}

\item{contrast}{A three element vector specifying the two samples to be compared, e.g. contrast=c("condition","Brain","UHR").}
}
\value{
A tibble object containing APA dynamics metrics of all APA genes.
}
\description{
Quantify APA dynamics among canonical PACs of a specific gene, canonical PACs represent PACs located at 3' UTR.
}
