% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Quantify.CNCAPA}
\alias{Quantify.CNCAPA}
\title{Quantify the APA dynamics between canonical and non-canonical PACs in a gene.}
\usage{
Quantify.CNCAPA(QpolyA,colData,contrast)
}
\arguments{
\item{QpolyA}{A QuantifyPolyA object generated after poly(A) site clustering and annotation.}

\item{colData}{A data.frame object containing the information of experimental design.}

\item{contrast}{A two element vector specifying the two samples to be compared, e.g. contrast=c("condition","Brain","UHR").}
}
\value{
A tibble object containing APA dynamics metrics of all APA genes.
}
\description{
Quantify the APA dynamics between canonical and non-canonical PACs in a gene.
}
