% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Quantify.SplitAPA}
\alias{Quantify.SplitAPA}
\title{Quantify APA dynamics among split PACs}
\usage{
Quantify.SplitAPA(QpolyA,colData,contrast)
}
\arguments{
\item{QpolyA}{A QuantifyPolyA object generated after poly(A) site clustering and annotation.}

\item{colData}{A data.frame object containing the information of experimental design.}

\item{contrast}{A three element vector specifying the two samples to be compared, e.g. contrast=c("condition","Brain","UHR").}
}
\value{
A tibble object containing APA dynamics metrics of all split PACs.
}
\description{
Quantify APA dynamics among split PACs identified by the weighted density peak clustering.
}
