% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\docType{class}
\name{QuantifyPolyA-class}
\alias{QuantifyPolyA-class}
\title{Class QuantifyPolyA.}
\description{
Class \code{QuantifyPolyA.} defines a poly(A) dataset.
}
\section{Slots}{

\describe{
\item{\code{sample_names}}{A character vector storing names of samples.}

\item{\code{pre.polyA}}{A list storing un-clustered poly(A) sites.}

\item{\code{polyA}}{A data.frame storing PACs (Poly(A) Site Clusters) generated from the weighted density peak clustering.}

\item{\code{simple.clusters}}{A data.frame storing PACs generated from the k-nt iteratively clustering.}

\item{\code{split.clusters}}{A data.frame storing only the split PACs.}

\item{\code{tail_lengths}}{A list storing poly(A) tail length information for each sample.}

\item{\code{cluster_tail_lengths}}{A list storing poly(A) tail length information mapped to clusters for each sample.}
}}

