% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Remove.IP}
\alias{Remove.IP}
\title{Remove IP}
\usage{
Remove.IP(QpolyA, fasta, flank_len = 15, win_size = 10, min_A = 8)
}
\arguments{
\item{QpolyA}{A QuantifyPolyA object containing all raw poly(A) sites.}

\item{fasta}{A string specifying the location and name of the corresponding genome file in FASTA format.}

\item{flank_len}{An integer specifying the neighborhood to search for template polyA.}

\item{win_size}{An integer specifying the size of the sliding window.}

\item{min_A}{An integer specifying the minimum number of base A in a window.}
}
\value{
A QuantifyPolyA object containing poly(A) sites with IP removed.
}
\description{
Remove poly(A) sites which potentially result from internal priming events.
}
