% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Save.TailLengths}
\alias{Save.TailLengths}
\title{Save tail length information}
\usage{
Save.TailLengths(QpolyA, output_dir)
}
\arguments{
\item{QpolyA}{A QuantifyPolyA object containing tail length information.}

\item{output_dir}{A string specifying the directory to save tail length information.}
}
\description{
Save poly(A) tail length information for each sample separately
}
