% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{show,QuantifyPolyA-method}
\alias{show,QuantifyPolyA-method}
\alias{show,ANY-method}
\title{Method show.}
\usage{
\S4method{show}{QuantifyPolyA}(object)
}
\arguments{
\item{object}{A QuantifyPolyA object.}
}
\description{
Method show.
}
