% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{simpleCluster}
\alias{simpleCluster}
\title{Cluster poly(A) sites into groups by a certain distance iterative.}
\usage{
simpleCluster(points.gr, max.gapwidth = 24)
}
\arguments{
\item{points.gr}{A GRanges object constructed from poly(A) sites.}

\item{max.gapwidth}{A cutoff to limit the maximum distance between two adjacent sites in a PAC.}
}
\value{
A reduced GRanges object.
}
\description{
Cluster poly(A) sites into groups by a certain distance iterative.
}
