% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{summarize_mRNA_level_results_multi}
\alias{summarize_mRNA_level_results_multi}
\title{Generate summary of mRNA-level statistical results for multiple treatments}
\usage{
summarize_mRNA_level_results_multi(results, alpha = 0.05, control_group = "NC")
}
\description{
Generate summary of mRNA-level statistical results for multiple treatments
}
