% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\docType{package}
\name{APPLE-package}
\alias{APPLE}
\alias{APPLE-package}
\title{APPLE: poly(A) sites and tail-length analysis}
\description{
Align reads, extract poly(A) sites and tails, remove internal priming,
cluster and annotate sites, and compare tail lengths and APA usage.
}
\seealso{
\code{\link[=minimap2]{minimap2()}}, \code{\link[=Extract_polyAsite]{Extract_polyAsite()}}, \code{\link[=Load.PolyA]{Load.PolyA()}},
\code{\link[=Remove.IP]{Remove.IP()}}, \code{\link[=Cluster.PolyA]{Cluster.PolyA()}}, \code{\link[=Annotate.PolyA]{Annotate.PolyA()}}, \code{\link[=Filter.PolyA]{Filter.PolyA()}},
\code{\link[=Map.Tail]{Map.Tail()}}, \code{\link[=Tail.PCA]{Tail.PCA()}}, \code{\link[=Tail.DiffPair]{Tail.DiffPair()}}, \code{\link[=Quantify.GeneAPA]{Quantify.GeneAPA()}}
}
\author{
\strong{Maintainer}: Zi Wang \email{wangzi0406@qq.com}

Authors:
\itemize{
  \item Zi Wang \email{wangzi0406@qq.com} (College of the Environment and Ecology, Xiamen University)
}

}

