% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{buildGenomicRanges}
\alias{buildGenomicRanges}
\title{Construct genomic ranges}
\usage{
buildGenomicRanges(seqname, position, score, strand = "*", five_prime_end)
}
\arguments{
\item{seqname}{A vector containing the sequence (Chromosomes/Contigs) names of poly(A) sites.}

\item{position}{A numeric vector containing the genomic positions of poly(A) sites.}

\item{score}{A numeric vector containing the numbers of reads supporting each poly(A) site.}

\item{strand}{A character vector containing the strand information of poly(A) sites.}

\item{five_prime_end}{A numeric vector of read 5-prime coordinates.}
}
\value{
A Granges object.
}
\description{
Construct genomic ranges
}
