% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Plot.DEAPACounts}
\alias{Plot.DEAPACounts}
\title{Plot counts of differential APA events}
\usage{
Plot.DEAPACounts(
  result,
  level = c("gene", "PAS"),
  effect_cutoff = 0.1,
  adjusted_p_cutoff = 0.05
)
}
\arguments{
\item{result}{Gene- or PAS-level APA result table.}

\item{level}{Either \code{"gene"} for Distal/Proximal genes or \code{"PAS"} for Up/Down PASs.}

\item{effect_cutoff}{Absolute PD or delta-PSU cutoff.}

\item{adjusted_p_cutoff}{Adjusted P-value cutoff.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot counts of differential APA events
}

