% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Plot.DEAPAVolcano}
\alias{Plot.DEAPAVolcano}
\title{Plot gene-level differential APA}
\usage{
Plot.DEAPAVolcano(
  result,
  pd_cutoff = 0.1,
  q_cutoff = 0.05,
  facet_columns = NULL
)
}
\arguments{
\item{result}{Gene-level APA result table.}

\item{pd_cutoff}{Absolute PD cutoff.}

\item{q_cutoff}{Adjusted P-value cutoff.}

\item{facet_columns}{Number of facet columns. By default, one per contrast.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot gene-level differential APA
}

