% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{Plot.PASPCA}
\alias{Plot.PASPCA}
\title{Plot PAS-expression PCA}
\usage{
Plot.PASPCA(x)
}
\arguments{
\item{x}{A result returned by \code{\link[=DESeq2.PolyA]{DESeq2.PolyA()}} or a DESeqDataSet.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot PAS-expression PCA
}
