% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APPLE.R
\name{Plot.PASVolcano}
\alias{Plot.PASVolcano}
\title{Plot differential PAS expression}
\usage{
Plot.PASVolcano(
  result,
  treatment_group,
  control_group = "NC",
  padj_cutoff = 0.05,
  lfc_cutoff = 1
)
}
\arguments{
\item{result}{A DESeq2 result table, typically returned by \code{lfcShrink()}.}

\item{treatment_group}{Treatment-group label.}

\item{control_group}{Control-group label.}

\item{padj_cutoff}{Adjusted P-value cutoff.}

\item{lfc_cutoff}{Absolute log2 fold-change cutoff.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot differential PAS expression
}

